TH.ReifySimple

Description

Utilities for reifying simplified datatype info. It omits details that aren't usually relevant to generating instances that work with the datatype. This makes it easier to use TH to derive instances.

The "Simple" in the module name refers to the simplicity of the datatypes, not the module itself, which exports quite a few things which are useful in some circumstance or another. I anticipate that the most common uses of this will be the following APIs:

Getting info about a data or newtype declaration, via DataType, reifyDataType, and DataCon. This is useful for writing something which generates declarations based on a datatype, one of the most common uses of Template Haskell.

Getting nicely structured info about a named type. See TypeInfo and reifyType. This does not yet support reifying typeclasses, primitive type constructors, or type variables (TyVarI).

Currently, this module supports reifying simplified versions of the following Info constructors:

TyConI with DataD and NewtypeD (becomes a DataType value)

FamilyI becomes a DataFamily or TypeFamily value.

DataConI becomes a DataCon value.

In the future it will hopefully also have support for the remaining Info constructors, ClassI, ClassOpI, PrimTyConI, VarI, and TyVarI.

Reifying simplified type info

data TypeInfo

reifyType :: Name -> Q TypeInfo

Reifies a Name as a TypeInfo, and calls fail if this doesn't work. Use reify with infoToType if you want to handle the failure case more gracefully.

This does not yet support reifying typeclasses, primitive type constructors, or type variables (TyVarI).

infoToType :: Info -> Q (Maybe TypeInfo)

Convert an Info into a TypeInfo if possible, and otherwise yield Nothing. Needs to run in Q so that

reifyTypeNoDataKinds :: Name -> Q (Maybe TypeInfo)

Reifies type info, but instead of yielding a LiftedDataConInfo, will instead yield Nothing.

infoToTypeNoDataKinds :: Info -> Maybe TypeInfo

Convert an 'Info into a TypeInfo if possible. If it's a data constructor, instead of yielding LiftedDataConInfo, it will instead yield Nothing.

Reifying simplified info for specific declaration varieties

Datatype info

data DataType

Simplified info about a DataD. Omits deriving, strictness, kind info, and whether it's data or newtype.

Constructors

DataType

Fields

dtName :: Name

dtTvs :: [Name]

dtCxt :: Cxt

dtCons :: [DataCon]

reifyDataType :: Name -> Q DataType

Reify the given data or newtype declaration, and yields its DataType representation.

infoToDataType :: Info -> Maybe DataType

Data constructor info

data DataCon

Simplified info about a Con. Omits deriving, strictness, and kind info. This is much nicer than consuming Con directly, because it unifies all the constructors into one.

Constructors

DataCon

Fields

dcName :: Name

dcTvs :: [Name]

dcCxt :: Cxt

dcFields :: [(Maybe Name, Type)]

reifyDataCon :: Name -> Q DataCon

Reify the given data constructor.

infoToDataCon :: Info -> Maybe DataCon

typeToDataCon :: Name -> Type -> DataCon

Creates a DataCon given the Name and Type of a data-constructor. Note that the result the function type is *not* checked to match the provided Name.

Data family info

data DataFamily

Simplified info about a data family. Omits deriving, strictness, and kind info.

Constructors

DataFamily

Fields

dfName :: Name

dfTvs :: [Name]

dfInsts :: [DataInst]

data DataInst

Simplified info about a data family instance. Omits deriving, strictness, and kind info.

Constructors

DataInst

Fields

diName :: Name

diCxt :: Cxt

diParams :: [Type]

diCons :: [DataCon]

reifyDataFamily :: Name -> Q DataFamily

Reify the given data family, and yield its DataFamily representation.

infoToDataFamily :: Info -> Maybe DataFamily

Type family info

data TypeFamily

Simplified info about a type family. Omits kind info and injectivity info.

Constructors

TypeFamily

Fields

tfName :: Name

tfTvs :: [Name]

tfInsts :: [TypeInst]

data TypeInst

Simplified info about a type family instance. Omits nothing.

Constructors

TypeInst

Fields

tiName :: Name

tiParams :: [Type]

tiType :: Type

reifyTypeFamily :: Name -> Q TypeFamily

Reify the given type family instance declaration, and yields its TypeInst representation.

infoToTypeFamily :: Info -> Maybe TypeFamily

Other utilities

conToDataCons :: Con -> [DataCon]

Convert a Con to a list of DataCon. The result is a list because GadtC and RecGadtC can define multiple constructors.

reifyDataTypeSubstituted :: Type -> Q DataType

Like reifyDataType, but takes a Type instead of just the Name of the datatype. It expects a normal datatype argument (see typeToNamedCon).